third pass
