+++ frontend/src/lib/components/Playlist/Playlist.tsx
@@ -1,7 +1,7 @@
 import './Playlist.scss'
 
 import { IconCollapse } from '@posthog/icons'
-import { LemonButton, LemonButtonProps, LemonCollapse, LemonSelect, LemonSkeleton, Tooltip } from '@posthog/lemon-ui'
+import { LemonButton, LemonButtonProps, LemonCollapse, LemonSkeleton, Tooltip } from '@posthog/lemon-ui'
 import clsx from 'clsx'
 import { useResizeBreakpoints } from 'lib/hooks/useResizeObserver'
 import { IconChevronRight } from 'lib/lemon-ui/icons'
@@ -9,8 +9,6 @@ import { LemonTableLoader } from 'lib/lemon-ui/LemonTable/LemonTableLoader'
 import { range } from 'lib/utils'
 import { useEffect, useRef, useState } from 'react'
 import { DraggableToNotebook } from 'scenes/notebooks/AddToNotebook/DraggableToNotebook'
-
-import { RecordingUniversalFilters } from '~/types'
 
 import { Resizer } from '../Resizer/Resizer'
 
@@ -46,8 +44,7 @@ export type PlaylistProps<T> = {
     onChangeSections?: (activeKeys: string[]) => void
     'data-attr'?: string
     activeItemId?: string
-    filters?: RecordingUniversalFilters
-    setFilters?: (filters: Partial<RecordingUniversalFilters>) => void
+    controls?: JSX.Element | null
 }
 
 const CounterBadge = ({ children }: { children: React.ReactNode }): JSX.Element => (
@@ -74,8 +71,7 @@ export function Playlist<
     onSelect,
     onChangeSections,
     'data-attr': dataAttr,
-    filters,
-    setFilters,
+    controls,
 }: PlaylistProps<T>): JSX.Element {
     const [controlledActiveItemId, setControlledActiveItemId] = useState<T['id'] | null>(
         selectInitialItem && sections[0].items[0] ? sections[0].items[0].id : null
@@ -121,8 +117,7 @@ export function Playlist<
                         setActiveItemId={onChangeActiveItem}
                         onChangeSections={onChangeSections}
                         emptyState={listEmptyState}
-                        filters={filters}
-                        setFilters={setFilters}
+                        controls={controls}
                     />
                 )}
                 <Resizer
@@ -145,70 +140,6 @@ const CollapsedList = ({ onClickOpen }: { onClickOpen: () => void }): JSX.Elemen
     </div>
 )
 
-function SortedBy({
-    filters,
-    setFilters,
-}: {
-    filters?: RecordingUniversalFilters
-    setFilters?: (filters: Partial<RecordingUniversalFilters>) => void
-}): JSX.Element | null {
-    return filters && setFilters ? (
-        <div className="px-2 justify-end flex flex-row gap-2 w-full items-center">
-            <span className="font-medium">sorted by</span>
-            <LemonSelect
-                allowClear={false}
-                options={[
-                    {
-                        value: 'start_time',
-                        label: 'Latest',
-                    },
-                    {
-                        label: 'Longest',
-                        options: [
-                            {
-                                value: 'duration',
-                                label: 'Total duration',
-                            },
-                            {
-                                value: 'active_seconds',
-                                label: 'Active duration',
-                            },
-                            {
-                                value: 'inactive_seconds',
-                                label: 'Inactive duration',
-                            },
-                        ],
-                    },
-                    {
-                        label: 'Most active',
-                        options: [
-                            {
-                                value: 'click_count',
-                                label: 'Clicks',
-                            },
-                            {
-                                value: 'keypress_count',
-                                label: 'Key presses',
-                            },
-                            {
-                                value: 'mouse_activity_count',
-                                label: 'Mouse activity',
-                            },
-                        ],
-                    },
-                    {
-                        value: 'console_error_count',
-                        label: 'Most errors',
-                    },
-                ]}
-                size="small"
-                value={filters.order}
-                onChange={(order) => setFilters({ order })}
-            />
-        </div>
-    ) : null
-}
-
 function List<
     T extends {
         id: string | number
@@ -226,8 +157,7 @@ function List<
     onScrollListEdge,
     loading,
     emptyState,
-    filters,
-    setFilters,
+    controls,
 }: {
     title: PlaylistProps<T>['title']
     notebooksHref: PlaylistProps<T>['notebooksHref']
@@ -240,8 +170,7 @@ function List<
     onScrollListEdge: PlaylistProps<T>['onScrollListEdge']
     loading: PlaylistProps<T>['loading']
     emptyState: PlaylistProps<T>['listEmptyState']
-    filters?: RecordingUniversalFilters
-    setFilters?: (filters: Partial<RecordingUniversalFilters>) => void
+    controls?: JSX.Element | null
 }): JSX.Element {
     const [activeHeaderActionKey, setActiveHeaderActionKey] = useState<string | null>(null)
     const lastScrollPositionRef = useRef(0)
@@ -321,7 +250,7 @@ function List<
                             </LemonButton>
                         ))}
                     </div>
-                    <SortedBy filters={filters} setFilters={setFilters} />
+                    <div className="w-full border-b">{controls}</div>
                     <LemonTableLoader loading={loading} />
                 </div>
             </DraggableToNotebook>
